chore: move method into lib folder
